feat: migrate to easy-typed-intl (Goals components)
