+++ src/components/GoalCreateForm/GoalCreateForm.i18n/en.json
@@ -1,0 +1,9 @@
+{
+    "Something went wrong 😿": "Something went wrong 😿",
+    "We are creating new goal": "We are creating new goal...",
+    "Voila! Goal is here 🎉": "Voila! Goal is here 🎉",
+    "New goal": "New goal",
+    "Pro tip!": "Pro tip!",
+    "Press key to create the goal": "Press {key} to create the goal.",
+    "Create goal": "Create goal"
+}
