+++ src/components/GoalCreateForm/GoalCreateForm.i18n/ru.json
@@ -1,0 +1,9 @@
+{
+    "Something went wrong 😿": "Упс! Что-то пошло не так 😿",
+    "We are creating new goal": "Погоди немного, создаем цель...",
+    "Voila! Goal is here 🎉": "Ура! Цель готова 🎉",
+    "New goal": "Новая цель",
+    "Pro tip!": "Держи совет!",
+    "Press key to create the goal": "Нажми {key}, чтобы создать цель.",
+    "Create goal": "Создать цель"
+}
