+++ src/components/GoalDeleteModal/GoalDeleteModal.i18n/en.json
@@ -1,0 +1,6 @@
+{
+    "You are trying to delete goal": "You are trying to delete goal",
+    "To confirm deleting goal please type goal key below": "To confirm deleting goal {goal} please type goal key below.",
+    "Cancel": "Cancel",
+    "Yes, delete it": "Yes, delete it"
+}
