+++ src/components/GoalDeleteModal/GoalDeleteModal.i18n/index.ts
