+++ src/components/GoalDeleteModal/GoalDeleteModal.i18n/ru.json
@@ -1,0 +1,6 @@
+{
+    "You are trying to delete goal": "Вы пытаетесь удалить цель",
+    "To confirm deleting goal please type goal key below": "Чтобы подтвердить удаление цели {goal} введите ключ цели в поле ниже.",
+    "Cancel": "Отмена",
+    "Yes, delete it": "Удалить цель"
+}
