+++ src/components/GoalDeleteModal/GoalDeleteModal.tsx
@@ -1,6 +1,5 @@
 import React, { ChangeEvent, useCallback, useState } from 'react';
 import dynamic from 'next/dynamic';
-import { useTranslations } from 'next-intl';
 import { danger0 } from '@taskany/colors';
 import {
     Button,
@@ -14,9 +13,11 @@ import {
     ModalHeader,
 } from '@taskany/bricks';
 
-import { dispatchModalEvent, ModalEvent } from '../utils/dispatchModal';
+import { dispatchModalEvent, ModalEvent } from '../../utils/dispatchModal';
 
-const ModalOnEvent = dynamic(() => import('./ModalOnEvent'));
+import { tr } from './GoalDeleteModal.i18n';
+
+const ModalOnEvent = dynamic(() => import('../ModalOnEvent'));
 
 interface GoalDeleteModalProps {
     id: string;
@@ -26,8 +27,6 @@ interface GoalDeleteModalProps {
 }
 
 export const GoalDeleteModal: React.FC<GoalDeleteModalProps> = ({ id, onConfirm, onCancel }) => {
-    const t = useTranslations('GoalDeleteModal');
-
     const [deleteConfirmation, setDeleteConfirmation] = useState('');
 
     const onConfirmationInputChange = useCallback((e: ChangeEvent<HTMLInputElement>) => {
@@ -43,13 +42,13 @@ export const GoalDeleteModal: React.FC<GoalDeleteModalProps> = ({ id, onConfirm,
     return (
         <ModalOnEvent view="danger" event={ModalEvent.GoalDeleteModal}>
             <ModalHeader>
-                <FormTitle color={danger0}>{t('You are trying to delete goal')}</FormTitle>
+                <FormTitle color={danger0}>{tr('You are trying to delete goal')}</FormTitle>
             </ModalHeader>
 
             <ModalContent>
                 <Text>
-                    {t.rich('To confirm deleting goal please type goal key below', {
-                        goal: () => <b>{id}</b>,
+                    {tr.raw('To confirm deleting goal please type goal key below', {
+                        goal: <b>{id}</b>,
                     })}
                 </Text>
 
@@ -61,13 +60,13 @@ export const GoalDeleteModal: React.FC<GoalDeleteModalProps> = ({ id, onConfirm,
                     <FormActions flat="top">
                         <FormAction left />
                         <FormAction right inline>
-                            <Button size="m" text={t('Cancel')} onClick={onDeleteCancel} />
+                            <Button size="m" text={tr('Cancel')} onClick={onDeleteCancel} />
                             <Button
                                 size="m"
                                 view="danger"
                                 disabled={deleteConfirmation !== id}
                                 onClick={onConfirm}
-                                text={t('Yes, delete it')}
+                                text={tr('Yes, delete it')}
                             />
                         </FormAction>
                     </FormActions>
