+++ src/components/GoalEditForm/GoalEditForm.i18n/en.json
@@ -1,0 +1,7 @@
+{
+    "Something went wrong 😿": "Something went wrong 😿",
+    "We are saving your goal": "We are saving your goal...",
+    "Voila! Saved successfully 🎉": "Voila! Saved successfully 🎉",
+    "Edit the goal": "Edit the goal",
+    "Submit": "Save goal"
+}
