+++ src/components/GoalEditForm/GoalEditForm.i18n/ru.json
@@ -1,0 +1,7 @@
+{
+    "Something went wrong 😿": "Упс! Что-то пошло не так 😿",
+    "We are saving your goal": "Погоди немного, сохраняем цель...",
+    "Voila! Saved successfully 🎉": "Ура! Сохранили цель 🎉",
+    "Submit": "Сохранить",
+    "Edit the goal": "Редактирование цели"
+}
