+++ src/components/GoalForm/GoalForm.i18n/en.json
@@ -1,0 +1,20 @@
+{
+    "Goal's title": "Goal's title",
+    "And its description": "And its description",
+    "Goal's title is required": "Goal's title is required",
+    "Goal's title must be a string": "Goal's title must be a string",
+    "Goal's title must be longer than 10 symbols": "Goal's title must be longer than 10 symbols",
+    "Goal's description must be longer than 10 symbols": "Goal's description must be longer than 10 symbols",
+    "Goal's description is required": "Goal's description is required",
+    "Goal's description must be a string": "Goal's description must be a string",
+    "Goal's project or team are required": "Goal's project or team are required",
+    "Assign": "Assign",
+    "Enter name or email": "Name/Email",
+    "Enter project or team title": "Team/Project",
+    "Schedule": "Schedule",
+    "Date input mask placeholder": "mm/dd/yyyy",
+    "Date input mask": "99/99/9999",
+    "State": "State",
+    "Enter tag title": "Enter tag title",
+    "Priority": "Priority"
+}
