+++ src/components/GoalForm/GoalForm.i18n/ru.json
@@ -1,0 +1,20 @@
+{
+    "Goal's title is required": "Название цели обязательно",
+    "Goal's title must be a string": "Название цели должно быть строкой",
+    "Goal's title must be longer than 10 symbols": "Название цели должно быть длинее 10х символов",
+    "Goal's description must be longer than 10 symbols": "Описание цели должно быть длинее 10х символов",
+    "Goal's description is required": "Описание цели обязательно",
+    "Goal's description must be a string": "Описание цели должно быть строкой",
+    "Goal's project or team are required": "Проект или команда обязательны",
+    "Assign": "Ответственный",
+    "Enter name or email": "Имя/Почта",
+    "Enter project or team title": "Команда/Проект",
+    "Schedule": "Срок",
+    "Date input mask placeholder": "dd.mm.yyyy",
+    "Date input mask": "99.99.9999",
+    "State": "Статус",
+    "Enter tag title": "Название тэга",
+    "Goal's title": "Название цели",
+    "And its description": "Можно и описание, пригодится ведь",
+    "Priority": "Приоритет"
+}
