+++ src/components/GoalListItem/GoalListItem.i18n/en.json
@@ -1,0 +1,3 @@
+{
+    "by": "by"
+}
