+++ src/components/GoalListItem/GoalListItem.i18n/index.ts
