+++ src/components/GoalListItem/GoalListItem.i18n/ru.json
@@ -1,0 +1,3 @@
+{
+    "by": null
+}
