+++ src/components/GoalPreview/GoalPreview.i18n/en.json
@@ -1,0 +1,7 @@
+{
+    "Something went wrong 😿": "Something went wrong 😿",
+    "We are deleting the goal": "We are deleting the goal...",
+    "Deleted successfully 🎉": "Deleted successfully 🎉",
+    "Edit": "Edit",
+    "Delete": "Delete"
+}
