+++ src/components/GoalPreview/GoalPreview.i18n/index.ts
