+++ src/components/GoalPreview/GoalPreview.i18n/ru.json
@@ -1,0 +1,7 @@
+{
+    "Something went wrong 😿": "Упс! Что-то пошло не так 😿",
+    "We are deleting the goal": "Удаляем цель...",
+    "Deleted successfully 🎉": "Удалено 🎉",
+    "Edit": "Изменить",
+    "Delete": "Удалить"
+}
